+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -148,7 +148,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T22:03:39.496Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-06T09:45:53.820Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")